one minor issue with loading text
