+++ public/index.html
@@ -11,12 +11,7 @@
     <!-- <link rel="icon" type="image/png" href="images/favicon.png"> -->
     <link rel="stylesheet" href="css/styles.css">
     <link rel="stylesheet" href="css/draggable.css">
-    <!-- makes it possible to display brush. This temporary and will be replaced once computerlab figures this out-->
-    <!-- <meta http-equiv="refresh" content="0; url=./whiteboard/index.html" /> -->
-    <meta http-equiv="refresh" content="0; url=./whiteboard/sound.html" />
     <script src="../js/libs/p5.min.js"></script>
-    <script language="javascript" src="https://cdn.jsdelivr.net/npm/p5@1.3.1/lib/addons/p5.sound.min.js"></script>
-    <script language="javascript" type="text/javascript" src="js/sound.js"></script>
   </head>
   <body>
