+++ public/whiteboard/index.html
@@ -14,24 +14,21 @@
     <link rel="stylesheet" href="../css/draggable.css">
   </head>
   <body>
-
-
     <!-- <input type="range" min="5" max="100" step="1" value="50"
     onchange="brushSize(this.value)">
     <div id="sizeSlider">50px</div>
     <button id="saveButton" type="button">save canvas</button> -->
-
+    <!-- <img src = "https://images-cdn.9gag.com/photo/aDW7w2B_700b.jpg" >  -->
     <script src="../js/libs/p5.min.js"></script>
-<<<<<<< HEAD
-=======
     <script src="/socket.io/socket.io.js"></script>
+    <script src="https://www.youtube.com/iframe_api"></script>
+    <script src="js/YTSampler.js" ></script>
     <script src="js/Brush.js"></script>
     <script src="js/User.js"></script>    
     <script src="js/OwnUser.js"></script>    
     <script src="js/UserManager.js"></script>    
     <script src="js/main.js"></script>
-
-    
->>>>>>> 4447a0d437ea0bf6d13de52522ac690b693bb100
+ 
+  
   </body>
 </html>
