--- public/whiteboard/sound.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title> Tutorial Page</title>
-  </head>
-  <body>
-      <!-- <img src = "https://images-cdn.9gag.com/photo/aDW7w2B_700b.jpg" > -->
-      <yt-sampler></yt-sampler>
-      <script src="https://www.youtube.com/iframe_api"></script>
-      <script src="sound.js" ></script>
-  </body>
-</html>
